backend/requirements.txt
Django>=4.2,<5
djangorestframework
djangorestframework-simplejwt
gunicorn
Pillow
python-dotenv
django-cors-headers
whitenoise
